Move driver into separate module
