+++ src/mattermostautodriver/driver/__init__.py
@@ -1,0 +1,4 @@
+from .driver import Driver, AsyncDriver
+
+
+__all__ = ["Driver", "AsyncDriver"]
